requirements.txt
pdfplumber==0.11.7
fastapi==0.116.2
uvicorn==0.35.0
sentence-transformers==5.1.0
weaviate-client==4.16.10
python-dotenv==1.1.1
python-multipart==0.0.20
redis==6.4.0
storage==0.0.4.3
SQLAlchemy==2.0.43
pydantic[email]
accelerate==1.10.1